+++ CHANGELOG.md
@@ -14,6 +14,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Changed
 
+    - [cmd/init] create a commit message by default, can be disabled by option `--no-commit` (GitHub Issue #23)
     - [cmd/set] renamed option `--labels` to `--labels-add`, which added specific labels (GitHub Issue #26)
 
 ### Removed