+++ CHANGELOG.md
@@ -9,6 +9,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
+    - [cmd] added `--version` / `-V` command to print the version of git-issue
     - [cmd/set] added option `--labels` to overwrite all labels (GitHub Issue #26)
     - [cmd/set] added option `--labels-remove` to remove specific labels (GitHub Issue #26)